Creating the Calculator main window
The first step of the workflow to follow in order to create the calculator GUI is:

1.Define the Main Window. For this, I use the Calculator class that inherits from QMainWindow(). The super().__init()
method initializes the Parent Class constructor method. I also set a few window properties like the Window Title, its
fixed size and its icon. I had to keep in mind that my Main Window need to have a central widget, so I defined the
QWidget as the central widget