Some changes in the Calculator Class and Added a program controller (Controller class)
The first change was adding the 'Times'Font to the display and setting the text alignment to the right.
Another change was in the _createButtons dictionary, first I made an empty dict and added the button text as the key and the final button widget
as the value in that dictionary, so I can have access to the widgets that I recently made.

I included a big change. I made a Controller Class responsible of connecting the signals with another functions. If the button pressed is "CE", I
connect the signal with the clearDisplay function, if the button pressed by the user is not "CE" or "=", the signal connects with the 'BuildingText'
Function, this function receives the text button, gets the current display text and concatenate those strings. Then, calls the setDisplayText
function, so the concatenated text is displayed on screen.

PD: The controller class receives the Calculator window instance to get access to the dictionary that store all the button widgets.